New Additions
DiskItemCollection added.  It's a addition to the DiskItem tool, which allows for objects to be paired with a guid.

DiskItemLibrary added.  It's an addition to the DiskItemCollection which allows only one instance of an object to be 'borrowed' at a time.  That object must be returned to the library before another instance can be accessed.